Switch internal-domain-secret webhook
